fix delete address bug and add color column to ProductImages

+++ Application/Services/address_service.py
@@ -3,8 +3,10 @@ from uuid import UUID
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from Application.Database.repositories import address_repository
-from Domain.Errors.address import AddressLimit, InvalidProvince, NoAddressWasFound
+from Application.Database.repositories import order_repository
+from Domain.Errors.address import AddressLimit, InvalidProvince, NoAddressWasFound, ActiveOrderAddress
 from Domain.schemas.address_schemas import NewAddress, UpdateAddress
+from utils import json_response
 
 IRANIAN_PROVINCES = {
     "البرز",
@@ -57,6 +59,9 @@ async def new_address(db: AsyncSession, address: NewAddress, user_id: UUID):
 
 
 async def delete_address(db: AsyncSession, user_id: UUID, address_id: int):
+    if await order_repository.check_pending_order_address(db = db, address_id = address_id, user_id = user_id):
+        raise ActiveOrderAddress
+    
     if await address_repository.delete_address(
         db=db, address_id=address_id, user_id=user_id
     ):
@@ -80,5 +85,8 @@ async def update_address(
 
 async def my_addresses(db: AsyncSession, user_id: UUID):
     Addresses = await address_repository.get_my_addresses(db, user_id=user_id)
+
+    if not Addresses:
+        return json_response(msg = "You have not registered an address yet.")
 
     return [i["Addresses"] for i in Addresses]
